postprocess values - sorting - DONE

+++ Inc/buffer.h
@@ -12,7 +12,7 @@
 
 typedef struct
 {
-	uint32_t size;
+	uint16_t size;
 	volatile uint16_t w_idx;
 	volatile uint16_t r_idx;
 	volatile uint16_t* buf;
